backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv